NGINX agent cannot find the default network interface on Linux hosts. (#317)
The parser was skipping over the first (header) line with the first
invocation of scanner.Scan().  The second invocation of scanner.Scan()
called in the for loop points to the first line of route data; however,
this line was being skipped.  The tests were incorrect because they
started with a newline before the header.