Change environment variable list to table with CLI references
This commit changes the list of NGINX Agent variables to a table showing
how it maps to the CLI flags, with an exception for the variable that
doesn't have a flag yet. It links to the CLI flag section so that the
reader can view the descriptions for additional context.