Update environment var table to have descriptions instead of pairs
The previous iteration of this table mapped the environment variables
to the CLU flags: this one instead replicates the descriptions from the
flags, ensuring a user doesn't need to cross-reference them.

It also changes the table to an include, meaning it can be easily
updated and re-used from a single source in the future if necessary.