remove concurrency in UploadMultipleFiles
Streams are not concurrency-safe to Send simultaneously.

We could potentially do something here with having several clients
each uploading a portion of the files but since Oliver was telling me
the point of the benchmark was to compare sequential stream vs
parallel unary I tried to just make it do that.

With this the 1000 files benchmark completes reasonably well for the
10k lines and under case.