Update linter version and install instructions in README.md (#1081)
* docs: fix mdatagen version in README

* chore: fix markdownlint errors in README

The only visible change is that the mdatagen issue is now linked. All
the rest of these are just markdownlint recommendations.

* test: fix assert reference to current test

This should fail the local inner test, not the parent.

* test: disable test host's OsReleaseInfo

`make unit-test` was failing on tests for parsing hostinfo. E.g.

- VersionId: (string) (len=5) "22.04",
- Version: (string) (len=29) "22.04.5 LTS (Jammy Jellyfish)",
+ VersionId: (string) (len=5) "24.04",
+ Version: (string) (len=26) "24.04.2 LTS (Noble Numbat)",

My host system (where make is running) is Ubuntu 24.04.2 LTS (Noble
Numbat) so it picking that up.

After tracing this through I found that

- The test is mocking what releaseinfo should return: internal/datasource/host/info_test.go#L518
- mock is consumed at internal/datasource/host/info.go#L298
- but then, internal/datasource/host/info.go#L299-L306
  - readOsRelease info (which is different than the "host"?)
     - if this errors returns the mocked hostReleaseInfo, e.g. on mac
     - if this returns data (my ubuntu host) it is merged into the returned data

It looks like these tests generally haven't been run on a host that
has `/etc/os-release` and tests shouldn't be subject to host system
data changes like that.

I've just put in a bad path so that on ubuntu hosts it won't read
anything; this should match the existing behavior of running this on
macos hosts.

* chore: Upgrade golangci-lint for compatibility with go1.24

When the local go toolchain is go1.24 or newer make lint reports a
bunch of spurious errors. This newer version of golangci-lint is
compatible with go1.24 eliminating that.

- `context` config changed, there are now more options than true or
false. "" disables it to match the previous behavior
- `exported-is-used` is deprecated, it is always true now.