Allow only CA cert to be set in command server TLS settings (#1116)
* test: use smaller selfsigned certs for testing

When generating a self-signed cert for _unit-tests_ it doesn't need to
have high security. Using a smaller length makes the tests run faster.

The difference is ~0.1s vs ~2.0s to run one test that generates a cert.

* test: fix grpc mTLS dialoptions test

This was incorrectly passing in the cert and key backwards which
resulted in not actually getting mTLS credentials, but instead
insecure credentials.

When insecure credentials are used, skipToken means we don't add a
addPerRPCCredentials. When it is a secure TLS credential then
addPerRPCCredentials increases the dialoption count by one.

* fix: Use specified CA cert for grpc

This had been skipping out of the function early if a client key
wasn't specified.

I don't believe that's correct. If I[User] have specified specified a
CA cert because the MPI server I'm trying to talk to is signed by a
non-standard CA (e.g. N1 devenv) then it should be respected
regardless of whether I've configured mTLS.

Silently skipping the CA is really confusing and leads to

> Failed to create connection" error="rpc error: code = Unavailable desc = connection error: desc = \"transport: authentication handshake failed: tls: failed to verify certificate: x509: certificate signed by unknown authority\"

I've split out the getTLSConfigForCredentials to make it easier to
test this translation. Once it is wrapped into a TransportCredential
or a DialOption it's opaque and hard to verify.

* fix: invalid TLS CA cert should error immediately

Previously if a consumer specified the CA cert to verify the command
connection but that CA wasn't valid then system would log at
Debug (default hidden) and proceed anyways.

I don't believe this is good behavior. If the consumer is directly
specifying a CA cert then that is the CA that should be used, not
silently ignored.

This patch returns the error up, which is now caught and swallowed at
a higher level, but at least it is more  visible:

> time=2025-05-21T15:41:33.547Z level=ERROR msg="Unable to add transport credentials to gRPC dial options, adding default transport credentials" error="invalid CA cert while building transport credentials: read CA file (/etc/nginx-agent/bad.crt): open /etc/nginx-agent/bad.crt: no such file or directory"

---------

Co-authored-by: Nathan Bird <n.bird@f5.com>